Possible fix for DB timeout
The backend runs continuously. As described in trello, after a week the fly proxy throws an timeout and after several minutes, the DB gets a timeout leading to a crash.

Now, when a connection is retreived from the singleton, we also check if we can still ping the server. Otherwise, a connection is re-established.

Though, a ping might still work in above situation or a re-connection might not be possible, we'll see next week.